requirements.txt
agno==2.0.11
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.69.0
anyio==4.11.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
decorator==5.2.1
distro==1.9.0
docstring_parser==0.17.0
elevenlabs @ git+https://github.com/elevenlabs/elevenlabs-python@1d86359ece25809e0c84b2b46c43f12c9fa5dc2d
exceptiongroup==1.3.0
executing==2.2.1
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipython==8.37.0
jedi==0.19.2
jiter==0.11.0
markdown-it-py==4.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.6.4
packaging==25.0
parso==0.8.5
pexpect==4.9.0
prompt_toolkit==3.0.52
propcache==0.3.2
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.4.2
pydantic-settings==2.11.0
pydantic_core==2.10.1
Pygments==2.19.2
PyJWT==2.10.1
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
rich==14.1.0
shellingham==1.5.4
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
traitlets==5.14.3
twilio==9.8.2
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.14
websockets==15.0.1
yarl==1.20.1
